Rename xeus_executor to python_executor
- Renamed pdit/xeus_executor.py → pdit/python_executor.py
- Renamed tests/test_xeus_executor.py → tests/test_python_executor.py
- Updated all imports to use PythonExecutor from python_executor
- Updated __init__.py to export PythonExecutor instead of XeusPythonExecutor
- File kept backwards compatibility alias for external users

The old name was misleading since we switched from xeus-python to
IPython kernel (python3).
